requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
cffi==1.14.4
chardet==3.0.4
discord.py==1.5.1
idna==2.10
multidict==4.7.6
pycparser==2.20
PyNaCl==1.4.0
six==1.15.0
yarl==1.5.1
